requirements.txt
Babel==2.6.0
beautifulsoup4==4.6.0
certifi==2019.6.16
chardet==3.0.4
deap==1.2.2
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0.3
django-extensions==2.1.9
django-import-export==1.2.0
django-modelcluster==4.4
django-phonenumber-field==2.2.0
django-taggit==0.23.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
draftjs-exporter==2.1.5
et-xmlfile==1.0.1
graphviz==0.11.1
html5lib==1.0.1
idna==2.8
jdcal==1.4
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.0
pandas==0.24.2
phonenumbers==8.10.5
Pillow==5.4.1
psycopg2==2.8.3
pydot==1.4.1
pydotplus==2.0.2
pygraphviz==1.5
pyparsing==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.3
PyYAML==3.13
reportlab==3.5.13
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sorl-thumbnail==12.5.0
stopit==1.1.2
tablib==0.12.1
TPOT==0.10.0
tqdm==4.31.1
unicodecsv==0.14.1
Unidecode==1.0.23
update-checker==0.16
urllib3==1.24.1
wagtail==2.4
webencodings==0.5.1
Willow==1.1
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
